pyproject.toml
[tool.poetry]
name = "cw-pyt-lib"
version = "0.1.0"
description = ""
authors = ["Michael Gorbunov <mg0rby@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
openpyxl = "^3.1.4"
xlrd = "^2.0.1"
requests = "^2.32.3"
pandas-stubs = "^2.2.2.240603"
types-requests = "^2.32.0.20240622"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"


[tool.poetry.group.dev.dependencies]
requests = "^2.32.3"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
#exclude =  ["^.venv/", "^tests/", "^di4/"]
exclude =  ["^.venv/", "^tests/", "^di4/"]
check_untyped_defs = true

[tool.black]
# Максимальная длина строки
line-length = 119
#exclude = [/(\.eggs|\.git|\.hg|\.mypy_cache|\.nox|\.tox|\.venv|_build|buck-out|build|dist)]
exclude =  ".venv/|di4/|.git|.mypy_cache"

[tool.isort]
# максимальная длина строки
line_length = 119
